move Library specific prefs and spinner initialization out to separate files
