+++ sampleapp/src/main/java/io/matthewnelson/sampleapp/ui/fragments/settings/library/SettingsLibraryFragment.kt
@@ -66,129 +66,16 @@
 * */
 package io.matthewnelson.sampleapp.ui.fragments.settings.library
 
-import android.content.Context
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.AdapterView
-import android.widget.ArrayAdapter
-import android.widget.Spinner
-import androidx.core.app.NotificationCompat
-import io.matthewnelson.encrypted_storage.Prefs
-import io.matthewnelson.sampleapp.App
-import io.matthewnelson.sampleapp.BuildConfig
 import io.matthewnelson.sampleapp.R
 
 class SettingsLibraryFragment : Fragment() {
 
-    companion object {
-        private val prefs: Prefs
-            get() = App.prefs
-
-        // Notification Keys
-        private const val NOTIFICATION_VISIBILITY = "NOTIFICATION_VISIBILITY"
-        private const val NOTIFICATION_COLOR_RESOURCE = "NOTIFICATION_COLOR_RESOURCE"
-        private const val NOTIFICATION_ENABLE_RESTART = "NOTIFICATION_ENABLE_RESTART"
-        private const val NOTIFICATION_ENABLE_STOP = "NOTIFICATION_ENABLE_STOP"
-        private const val NOTIFICATION_SHOW = "NOTIFICATION_SHOW"
-
-        fun getNotificationVisibilitySetting(): Int =
-            prefs.read(
-                NOTIFICATION_VISIBILITY, NotificationCompat.VISIBILITY_PRIVATE)
-
-        fun getNotificationColorSetting(): Int =
-            prefs.read(
-                NOTIFICATION_COLOR_RESOURCE, R.color.primaryColor)
-
-        fun getNotificationRestartEnableSetting(): Boolean =
-            prefs.read(
-                NOTIFICATION_ENABLE_RESTART, true)
-
-        fun getNotificationStopEnableSetting(): Boolean =
-            prefs.read(
-                NOTIFICATION_ENABLE_STOP, true)
-
-        fun getNotificationShowSetting(): Boolean =
-            prefs.read(
-                NOTIFICATION_SHOW, true)
-
-        // BackgroundManager Keys
-        private const val BACKGROUND_MANAGER_POLICY = "BACKGROUND_MANAGER_POLICY"
-        private const val BACKGROUND_MANAGER_EXECUTE_DELAY = "BACKGROUND_MANAGER_EXECUTE_DELAY"
-        private const val BACKGROUND_MANAGER_KILL_APP = "BACKGROUND_MANAGER_KILL_APP"
-
-        // values
-        const val BACKGROUND_MANAGER_POLICY_RESPECT = "BACKGROUND_MANAGER_POLICY_RESPECT"
-        const val BACKGROUND_MANAGER_POLICY_FOREGROUND = "BACKGROUND_MANAGER_POLICY_FOREGROUND"
-
-        fun getBackgroundManagerPolicySetting(): String? =
-            prefs.read(
-                BACKGROUND_MANAGER_POLICY, Prefs.INVALID_STRING)
-
-        fun getBackgroundManagerExecuteDelaySetting(): Int =
-            prefs.read(
-                BACKGROUND_MANAGER_EXECUTE_DELAY, 20)
-
-        fun getBackgroundManagerKillAppSetting(): Boolean =
-            prefs.read(
-                BACKGROUND_MANAGER_KILL_APP, true)
-
-        // Controller Keys
-        private const val CONTROLLER_BUILD_CONFIG_DEBUG = "CONTROLLER_BUILD_CONFIG_DEBUG"
-        private const val CONTROLLER_RESTART_DELAY = "CONTROLLER_RESTART_DELAY"
-        private const val CONTROLLER_STOP_DELAY = "CONTROLLER_STOP_DELAY"
-        private const val CONTROLLER_DISABLE_STOP_SERVICE_TASK_REMOVED = "CONTROLLER_DISABLE_STOP_SERVICE_TASK_REMOVED"
-
-        fun getControllerRestartDelaySetting(): Long =
-            prefs.read(
-                CONTROLLER_RESTART_DELAY, 100L)
-
-        fun getControllerStopDelaySetting(): Long =
-            prefs.read(
-                CONTROLLER_STOP_DELAY, 100L)
-
-        fun getControllerDisableStopServiceOnTaskRemovedSetting(): Boolean =
-            prefs.read(
-                CONTROLLER_DISABLE_STOP_SERVICE_TASK_REMOVED, false)
-
-        fun getControllerBuildConfigDebugSetting(): Boolean =
-            prefs.read(
-                CONTROLLER_BUILD_CONFIG_DEBUG, BuildConfig.DEBUG)
-
-
-        // Spinner Notification Visibility
-        private const val PRIVATE = "Private"
-        private const val PUBLIC = "Public"
-        private const val SECRET = "Secret"
-
-        // Spinner Notification Color
-        private const val NONE = "None"
-        private const val TOR_PURPLE = "Tor Purple"
-        private const val YELLOW = "Yellow"
-        private const val GREEN = "Green"
-        private const val BLUE = "Blue"
-
-        // Spinner Notification Buttons
-        private const val ENABLED = "Enabled"
-        private const val DISABLED = "Disabled"
-
-        // Spinner Notification Show
-        private const val SHOW = "Show"
-        private const val HIDE = "Hide"
-    }
-
-    private lateinit var spinnerVisibility: Spinner
-    private lateinit var adapterVisibility: ArrayAdapter<String>
-    private lateinit var spinnerColor: Spinner
-    private lateinit var adapterColor: ArrayAdapter<String>
-    private lateinit var spinnerRestart: Spinner
-    private lateinit var adapterRestart: ArrayAdapter<String>
-    private lateinit var spinnerStop: Spinner
-    private lateinit var adapterStop: ArrayAdapter<String>
-    private lateinit var spinnerShow: Spinner
-    private lateinit var adapterShow: ArrayAdapter<String>
+    private lateinit var notificationSpinners: NotificationSpinners
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -200,262 +87,10 @@ class SettingsLibraryFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        findViews(view)
-        initSpinnerNotificationVisibility(view.context)
-        initSpinnerNotificationColor(view.context)
-        initSpinnerNotificationRestartButton(view.context)
-        initSpinnerNotificationStopButton(view.context)
-        initSpinnerNotificationShow(view.context)
+        notificationSpinners = NotificationSpinners(view)
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
-    }
-
-    private fun findViews(view: View) {
-        spinnerVisibility = view.findViewById(R.id.settings_library_spinner_notification_visibility)
-        spinnerColor = view.findViewById(R.id.settings_library_spinner_notification_icon_color)
-        spinnerRestart = view.findViewById(R.id.settings_library_spinner_notification_restart_button)
-        spinnerStop = view.findViewById(R.id.settings_library_spinner_notification_stop_button)
-        spinnerShow = view.findViewById(R.id.settings_library_spinner_notification_show)
-    }
-
-    private fun initSpinnerNotificationVisibility(context: Context) {
-        val categoryVisibility = arrayOf(
-            PRIVATE,
-            PUBLIC,
-            SECRET
-        )
-        adapterVisibility = ArrayAdapter(context, R.layout.spinner_list_item, categoryVisibility)
-        adapterVisibility.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-        spinnerVisibility.adapter = adapterVisibility
-        setNotificationVisibilitySpinnerValue()
-        spinnerVisibility.onItemSelectedListener = NotificationOptionSelectionListener()
-    }
-
-    private fun setNotificationVisibilitySpinnerValue() {
-        when (getNotificationVisibilitySetting()) {
-            NotificationCompat.VISIBILITY_PRIVATE -> {
-                spinnerVisibility.setSelection(0)
-            }
-            NotificationCompat.VISIBILITY_PUBLIC -> {
-                spinnerVisibility.setSelection(1)
-            }
-            NotificationCompat.VISIBILITY_SECRET -> {
-                spinnerVisibility.setSelection(2)
-            }
-        }
-    }
-
-    private fun initSpinnerNotificationColor(context: Context) {
-        val categoryColor = arrayOf(
-            NONE,
-            TOR_PURPLE,
-            YELLOW,
-            GREEN,
-            BLUE
-        )
-        adapterColor = ArrayAdapter(context, R.layout.spinner_list_item, categoryColor)
-        adapterColor.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-        spinnerColor.adapter = adapterColor
-        setNotificationColorSpinnerValue()
-        spinnerColor.onItemSelectedListener = NotificationOptionSelectionListener()
-    }
-
-    private fun setNotificationColorSpinnerValue() {
-        when (getNotificationColorSetting()) {
-            R.color.tor_service_white -> {
-                spinnerColor.setSelection(0)
-            }
-            R.color.primaryColor -> {
-                spinnerColor.setSelection(1)
-            }
-            R.color.yellow -> {
-                spinnerColor.setSelection(2)
-            }
-            R.color.green -> {
-                spinnerColor.setSelection(3)
-            }
-            R.color.blue -> {
-                spinnerColor.setSelection(4)
-            }
-        }
-    }
-
-    private fun initSpinnerNotificationRestartButton(context: Context) {
-        val categoryEnabledDisabled = arrayOf(
-            ENABLED,
-            DISABLED
-        )
-        adapterRestart = ArrayAdapter(context, R.layout.spinner_list_item, categoryEnabledDisabled)
-        adapterRestart.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-        spinnerRestart.adapter = adapterRestart
-        setNotificationRestartButtonSpinnerValue()
-        spinnerRestart.onItemSelectedListener = NotificationOptionSelectionListener()
-    }
-
-    private fun setNotificationRestartButtonSpinnerValue() {
-        when (getNotificationRestartEnableSetting()) {
-            true -> {
-                spinnerRestart.setSelection(0)
-            }
-            false -> {
-                spinnerRestart.setSelection(1)
-            }
-        }
-    }
-
-    private fun initSpinnerNotificationStopButton(context: Context) {
-        val categoryEnabledDisabled = arrayOf(
-            ENABLED,
-            DISABLED
-        )
-        adapterStop = ArrayAdapter(context, R.layout.spinner_list_item, categoryEnabledDisabled)
-        adapterStop.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-        spinnerStop.adapter = adapterStop
-        setNotificationStopButtonSpinnerValue()
-        spinnerStop.onItemSelectedListener = NotificationOptionSelectionListener()
-    }
-
-    private fun setNotificationStopButtonSpinnerValue() {
-        when (getNotificationStopEnableSetting()) {
-            true -> {
-                spinnerStop.setSelection(0)
-            }
-            false -> {
-                spinnerStop.setSelection(1)
-            }
-        }
-    }
-
-    private fun initSpinnerNotificationShow(context: Context) {
-        val categoryEnabledDisabled = arrayOf(
-            SHOW,
-            HIDE
-        )
-        adapterShow = ArrayAdapter(context, R.layout.spinner_list_item, categoryEnabledDisabled)
-        adapterShow.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-        spinnerShow.adapter = adapterShow
-        setNotificationShowSpinnerValue()
-        spinnerShow.onItemSelectedListener = NotificationOptionSelectionListener()
-    }
-
-    private fun setNotificationShowSpinnerValue() {
-        when (getNotificationShowSetting()) {
-            true -> {
-                spinnerShow.setSelection(0)
-            }
-            false -> {
-                spinnerShow.setSelection(1)
-            }
-        }
-    }
-
-    private inner class NotificationOptionSelectionListener: AdapterView.OnItemSelectedListener {
-        var count = 0
-            private set
-
-        override fun onNothingSelected(parent: AdapterView<*>?) {}
-
-        override fun onItemSelected(parent: AdapterView<*>?, view: View?, position: Int, id: Long) {
-            if (count == 0) {
-                count++
-                return
-            }
-
-            val item = parent?.getItemAtPosition(position) ?: return
-
-            when (parent.adapter) {
-                adapterVisibility -> {
-                    val visibility: Int = when (item.toString()) {
-                        PRIVATE -> {
-                            NotificationCompat.VISIBILITY_PRIVATE
-                        }
-                        PUBLIC -> {
-                            NotificationCompat.VISIBILITY_PUBLIC
-                        }
-                        SECRET -> {
-                            NotificationCompat.VISIBILITY_SECRET
-                        }
-                        else -> {
-                            return
-                        }
-                    }
-                    prefs.write(
-                        NOTIFICATION_VISIBILITY, visibility)
-                }
-                adapterColor -> {
-                    val colorRes: Int = when (item.toString()) {
-                        NONE -> {
-                            R.color.tor_service_white
-                        }
-                        TOR_PURPLE -> {
-                            R.color.primaryColor
-                        }
-                        YELLOW -> {
-                            R.color.yellow
-                        }
-                        GREEN -> {
-                            R.color.green
-                        }
-                        BLUE -> {
-                            R.color.blue
-                        }
-                        else -> {
-                            return
-                        }
-                    }
-                    prefs.write(
-                        NOTIFICATION_COLOR_RESOURCE, colorRes)
-                }
-                adapterRestart -> {
-                    val enabled: Boolean = when (item.toString()) {
-                        ENABLED -> {
-                            true
-                        }
-                        DISABLED -> {
-                            false
-                        }
-                        else -> {
-                            return
-                        }
-                    }
-                    prefs.write(
-                        NOTIFICATION_ENABLE_RESTART, enabled)
-                }
-                adapterStop -> {
-                    val enabled: Boolean = when (item.toString()) {
-                        ENABLED -> {
-                            true
-                        }
-                        DISABLED -> {
-                            false
-                        }
-                        else -> {
-                            return
-                        }
-                    }
-                    prefs.write(
-                        NOTIFICATION_ENABLE_STOP, enabled)
-                }
-                adapterShow -> {
-                    val show: Boolean = when (item.toString()) {
-                        SHOW -> {
-                            true
-                        }
-                        HIDE -> {
-                            false
-                        }
-                        else -> {
-                            return
-                        }
-                    }
-                    prefs.write(
-                        NOTIFICATION_SHOW, show)
-                }
-            }
-            // TODO: add restart app popup
-        }
-
     }
 }
