+++ sampleapp/src/main/res/layout/fragment_settings_library.xml
@@ -192,7 +192,7 @@
                     android:layout_width="0dp"
                     android:layout_height="match_parent"
                     android:singleLine="true"
-                    android:text="@string/settings_library_text_view_notification_option_restart_button"
+                    android:text="@string/settings_library_text_view_notification_option_stop_button"
                     android:textStyle="bold"
                     app:layout_constraintBottom_toBottomOf="@+id/settings_library_spinner_notification_stop_button"
                     app:layout_constraintEnd_toEndOf="@+id/settings_library_notification_guide_left"
@@ -215,7 +215,7 @@
                     android:layout_width="0dp"
                     android:layout_height="match_parent"
                     android:singleLine="true"
-                    android:text="@string/settings_library_text_view_notification_option_restart_button"
+                    android:text="@string/settings_library_text_view_notification_option_show_notification"
                     android:textStyle="bold"
                     app:layout_constraintBottom_toBottomOf="@+id/settings_library_spinner_notification_show"
                     app:layout_constraintEnd_toEndOf="@+id/settings_library_notification_guide_left"
